tests/load/requirements.txt
locust>=2.29,<3
websocket-client>=1.7,<2
cryptography>=43,<45
psycopg2-binary>=2.9,<3
